pyproject.toml
[project]
name = "a2a-python-example-ui"
version = "0.1.0"
description = "Agent2Agent example UI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.10.6",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "mesop>=1.0.0",
    "a2a-sdk>=0.2.9",
    "pandas>=2.2.0",
    "google-genai>=1.9.0",
    "google-adk>=1.0.0",
    "click>=8.1.8",
    "fastmcp>=1.0",
    "google-cloud-aiplatform>=1.91.0",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.4.1",
    "mcp[cli]>=1.5.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "python-dotenv>=1.1.0",
    "marvin>=3.0.0",
    "emoji>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["a2a_ui"]

[tool.uv.sources]
a2a_ui = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.2",
] 
